backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai==1.3.6
python-dotenv==1.0.0
jsonschema==4.19.0
azure-identity==1.13.0
azure-mgmt-datafactory==3.1.0
azure-mgmt-resource==23.0.1
pytest==7.4.0
requests==2.31.0